:sparkles: add: added delete-ong use-cases

+++ src/modules/incidents/useCases/delete-incident/delete-incident-use-case.ts
@@ -1,9 +1,10 @@
-import { Injectable, Inject } from '@nestjs/common';
 import { IIncidentsRepository } from '@modules/incidents/repositories/incidents-repository.interface';
+import { Inject, Injectable } from '@nestjs/common';
 import { DeleteIncidentException } from './delete-incident-exception';
 
 type DeleteIncidentUseCaseRequest = {
   incidentId: string;
+  ongId?: string;
 };
 
 type DeleteIncidentUseCaseResponse = void;
@@ -17,11 +18,25 @@ export class DeleteIncidentUseCase {
 
   async execute({
     incidentId,
+    ongId,
   }: DeleteIncidentUseCaseRequest): Promise<DeleteIncidentUseCaseResponse> {
+    // const ongExists = await this.ongsRepository.findById();
     const incidentExists = await this.incidentsRepository.findById(incidentId);
 
     if (!incidentExists) {
       throw new DeleteIncidentException.NotFoundIncidentError();
+    }
+
+    // if (!ongExists.isAdmin) {
+    //   if (ongId !== incidentExists.ongId) {
+    //     throw new UnauthorizedException();
+    //   }
+
+    //   await this.incidentsRepository.delete(incidentExists);
+    // }
+
+    if (ongId !== incidentExists.ongId) {
+      throw new DeleteIncidentException.OngNotAllowedForbiddenException();
     }
 
     await this.incidentsRepository.delete(incidentExists);
